+++ src/test/kotlin/com/github/ekvedaras/classfactoryphpstorm/EssentialTestCase.kt
@@ -171,6 +171,13 @@ internal abstract class EssentialTestCase : TestCase() {
         )
     }
 
+    fun testItCanHandleClosureValues() {
+        assertInspection(
+            "essential/propertyWithClosureValue.php",
+            incorrectPropertyTypeInspection(),
+        )
+    }
+
     fun testItReportsIncorrectPropertyTypesWhenThereIsNoConstructor() {
         assertInspection(
             "essential/propertyWithWrongTypeWhenThereIsNoConstructor.php",